Don't trigger dock handle drag enter event unless a known item is being dragged
This seems a little counterintuitive. How could a dragenter event fire
if nothing's being dragged? Well, if something's dragged into the
document (i.e. from the desktop or another window), a dragenter event
will be fired even though (as far as we're concerned), nothing is being
dragged. Rather than trying to handle this class of events specially in
the dock, we simply limit our listening to when something is being
dragged. Since we have a single source of truth for this information,
this should be resilient to future changes in logic.